add timestamp and improve migration

+++ src/main/java/de/tum/in/www1/artemis/repository/competency/JolValueEntry.java
@@ -3,5 +3,5 @@ package de.tum.in.www1.artemis.repository.competency;
 /**
  * A simple value entry for JOL values.
  */
-public record JolValueEntry(long competencyId, int value) {
+public record JolValueEntry(long competencyId, short value) {
 }
